jwt authentication and authorization

+++ ConnectIO/controllers/api/v1/posts_api.js
@@ -34,12 +34,17 @@ module.exports.destroy = async function(req,res){
     try{
         let post = await Post.findById(req.params.id);
         //.id means converting object id into string. _id is ObjectID. 
-        post.remove();
-        await Comment.deleteMany({post: req.params.id});
+        
+        if(post.user == req.user.id){
+            post.remove();
+            await Comment.deleteMany({post: req.params.id});
 
-        return  res.json(200,{
-            message:"Post deleted successfully!"
-        });;
+            return  res.json(200,{message:"Post deleted successfully!"});;
+        }else{
+            return  res.json(401,{message:"UnAuthorized Access!!"});;
+        }
+        
+        
     }catch(err){
 
         console.log("Error : ",err);
